move gather_at_rank to utils
