+++ charcoal/just_taxonomy.py
@@ -21,7 +21,7 @@ from .lineage_db import LineageDB
 from .version import version
 from .utils import (get_idents_for_hashval, gather_lca_assignments,
     count_lca_for_assignments, pretty_print_lineage, pretty_print_lineage2,
-    WriteAndTrackFasta)
+    WriteAndTrackFasta, gather_at_rank, get_ident)
 
 
 GATHER_MIN_MATCHES=3
@@ -40,14 +40,6 @@ class ContigInfo(Enum):
 
 def kb(bp):
     return int(bp/1000)
-
-
-def get_ident(sig):
-    "Hack and slash identifiers."
-    ident = sig.name()
-    ident = ident.split()[0]
-    ident = ident.split('.')[0]
-    return ident
 
 
 class ContigReport(object):
@@ -132,39 +124,6 @@ def create_empty_output(genome, comment, summary, report, contig_report,
             pass
     open(clean, 'wt').close()
     open(dirty, 'wt').close()
-
-
-def gather_at_rank(mh, lca_db, lin_db, match_rank):
-    "Run gather, and aggregate at given rank."
-    import copy
-    minhash = copy.copy(mh)
-    query_sig = sourmash.SourmashSignature(minhash)
-
-    # do the gather:
-    counts = Counter()
-    while 1:
-        results = lca_db.gather(query_sig, threshold_bp=0)
-        if not results:
-            break
-
-        (match, match_sig, _) = results[0]
-
-        # retrieve lineage & pop to match_rank
-        match_ident = get_ident(match_sig)
-        match_lineage = lin_db.ident_to_lineage[match_ident]
-        match_lineage = utils.pop_to_rank(match_lineage, match_rank)
-
-        # count at match_rank
-        common = match_sig.minhash.count_common(query_sig.minhash)
-        counts[match_lineage] += common
-
-        # finish out gather algorithm!
-        minhash.remove_many(match_sig.minhash.get_mins())
-        query_sig = sourmash.SourmashSignature(minhash)
-
-    # return!
-    for lin, count in counts.most_common():
-        yield lin, count
 
 
 def guess_tax_by_gather(entire_mh, lca_db, lin_db, match_rank, report_fp):
@@ -577,8 +536,6 @@ def main(args):
                                           GATHER_MIN_MATCHES,
                                           genome_lineage, match_rank,
                                           report_fp)
-
-    print('XYZ', first_match)
 
     if not first_match:
         print(' ** no matches to clean sequence **', file=report_fp)
